# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: "Описание спецификации всех API методов, используемых в online diffusion\
    \ сервисе"
  title: Online diffusion API
  version: 0.0.1
  x-logo:
    url: https://i.imgur.com/1kUh580.jpg
...
servers:
- description: "Локальный сервер, актуально для запуска на своем компьютере"
  url: http://localhost:8080
  x-last-modified: 1698183338768
security:
- JWTAuth: []
...
paths:
...
  /api/v1/feed:
    get:
      description: ""
      operationId: feed
      parameters:
      - description: "Маркер, показывающий начиная с какого id подгружать ресурсы"
        explode: true
        in: query
        name: marker
        required: false
        schema:
          example: 1
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostPagingWrapper'
          description: "все хорошо, присланы посты"
        "404":
          description: такой маркер не найден
      security:
      - JWTAuth: []
      summary: Получение списка постов
      tags:
      - Images
      x-accepts: application/json
      x-tags:
      - tag: Images
  /api/v1/images/{photoId}:
    get:
      description: ""
      operationId: get-image
      parameters:
      - description: ID фото в базе данных
        explode: false
        in: path
        name: photoId
        required: true
        schema:
          example: 123456
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
          description: токен активирован и новый пароль успешно принят
        "400":
          description: "такой токен есть, однако он еще не был активирован присыла\
            нием правильного кода в методе Confirm_Password_Code"
        "404":
          description: такого recovery_token нет на сервере в принципе
      security:
      - JWTAuth: []
      summary: Получение полной информации про картинку
      tags:
      - Images
      x-accepts: application/json
      x-tags:
      - tag: Images
    post:
      description: ""
      operationId: post-image
      parameters:
      - description: ID фото в базе данных
        explode: false
        in: path
        name: photoId
        required: true
        schema:
          example: 123456
          format: int64
          type: integer
        style: simple
      responses:
        "201":
          description: изображение было успешно опубликовано
        "400":
          description: такого photo_id нет на сервере в принципе
        "404":
          description: "данное photo_id не может быть опубликовано пользователем,\
            \ так как это не его фотография или эта фотография уже была опубликова\
            на"
      security:
      - JWTAuth: []
      summary: Публикуем изображение
      tags:
      - Images
      x-accepts: application/json
      x-tags:
      - tag: Images
    put:
      description: ""
      operationId: put-image
      parameters:
      - description: ID фото в базе данных
        explode: false
        in: path
        name: photoId
        required: true
        schema:
          example: 123456
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: изображение было успешно сохранено в галерею
        "404":
          description: такого photo_id нет на сервере в принципе
        "409":
          description: "данное photo_id не может быть добавлено в галерею пользова\
            телем, так как это не его фотография или эта фотография уже есть у нег\
            о в галерее"
      security:
      - JWTAuth: []
      summary: Сохраняем в галерею
      tags:
      - Images
      x-accepts: application/json
      x-tags:
      - tag: Images
  /api/v1/images/make-avatar:
    put:
      description: ""
      operationId: make-avatar
      parameters:
      - description: ID фото в базе данных
        explode: true
        in: query
        name: targetPhotoId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content: {}
          description: изображение было успешно поставлено на автарку
        "404":
          description: такого photo_id нет на сервере в принципе
      security:
      - JWTAuth: []
      summary: Ставим изображение пользователю на аватарку
      tags:
      - Images
      x-accepts: application/json
      x-tags:
      - tag: Images
...
components:
...
  schemas:
    PostPagingWrapper:
      description: Оболочка для пагинации
      example:
        nextMarker: 0.8008281904610115
        posts:
        - null
        - null
      properties:
        nextMarker:
          description: Маркер для следующего запроса
          type: number
        posts:
          description: Список присланных постов в нужном порядке
          items:
            $ref: '#/components/schemas/Post'
          type: array
      type: object
...
    Post:
      description: Пост в ленте с изображением
      oneOf:
      - $ref: '#/components/schemas/Image'
      x-one-of-name: Post
...
    Image:
      description: Объект с информацией об фотографии
      example:
        seed: seed
        authorName: authorName
        authorAvatarUrl: authorAvatarUrl
        photoId: 0.8008281904610115
        anti-prompt: anti-prompt
        model: model
        prompt: prompt
      properties:
        photoId:
          description: Id фотографии
          type: number
        authorName:
          description: Имя автора
          type: string
        authorAvatarUrl:
          description: URL аватара автора
          type: string
        prompt:
          description: prompt использовшийся для создания изображения
          type: string
        anti-prompt:
          description: anti-prompt использовшийся для создания изображения
          type: string
        seed:
          description: "seed, использовшийся для создания изображения"
          type: string
        model:
          description: "Модель, использувшееся для создания изображения"
          type: string
      required:
      - anti-prompt
      - authorAvatarUrl
      - authorName
      - model
      - photoId
      - prompt
      - seed
      type: object
...
  securitySchemes:
    JWTAuth:
      bearerFormat: JWT
      description: "Example: \n> Authorization: Bearer <token>"
      scheme: bearer
      type: http
      x-last-modified: 1698183246392